+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/application/StatementController.kt
@@ -174,7 +174,8 @@ class StatementController(
         @RequestParam("minLevel", required = false) minLevel: Int?,
         @RequestParam("maxLevel", required = false) maxLevel: Int?,
         @RequestParam("blacklist", required = false, defaultValue = "") blacklist: List<ClassId>,
-        @RequestParam("whitelist", required = false, defaultValue = "") whitelist: List<ClassId>
+        @RequestParam("whitelist", required = false, defaultValue = "") whitelist: List<ClassId>,
+        @RequestParam("includeFirst", required = false, defaultValue = "true") includeFirst: Boolean
     ): HttpEntity<Bundle> {
         return ok(
             statementService.fetchAsBundle(
@@ -183,7 +184,8 @@ class StatementController(
                 BundleConfiguration(
                     minLevel, maxLevel,
                     blacklist, whitelist
-                )
+                ),
+                includeFirst
             )
         )
     }
@@ -226,14 +228,19 @@ data class BundleConfiguration(
         var labelFilter = ""
         if (blacklist.isNotEmpty())
             labelFilter = blacklist.joinToString(prefix = "-", separator = "|-")
-        if (blacklist.isNotEmpty()) {
+        if (whitelist.isNotEmpty()) {
             var positiveLabels = whitelist.joinToString(prefix = "+", separator = "|+")
             if (labelFilter.isNotBlank())
-                positiveLabels += "|$positiveLabels"
+                positiveLabels += "|$labelFilter"
             labelFilter = positiveLabels
         }
         if (blacklist.isNotEmpty() || whitelist.isNotEmpty())
             conf["labelFilter"] = labelFilter
         return conf
+    }
+
+    companion object Factory {
+        fun firstLevelConf(): BundleConfiguration =
+            BundleConfiguration(minLevel = null, maxLevel = 1, blacklist = emptyList(), whitelist = emptyList())
     }
 }